jailer: * Remove nix specific prepartion steps. * Move configuration of mount points to .fstab
nix-helper:
* Setup nix specific parts
* Write .fstab
* Add .Xauthority copying. Not really secure :/
* Add FHS chroot support and example.

shell-support:
* Add signature checking.

linker.c:
* Add recursive linker that helps with FHS setup.

+++ shell-support/shell-trigger.sh
@@ -1,16 +1,43 @@
-chpwd() {
-    SHELL_NIX=$(findup shell.nix)
-    if [ -n "$SHELL_NIX" ]; then
-	# The protocol between this helper and the shell spawned is that
-	# this shell will chdir to a dir written to file descriptor 3
-	# by nix-shell or its subprocesses.
-	CAGEFILE=$(mktemp /tmp/cage.XXXXXX)
-	exec 3> $CAGEFILE
-
-	nix-shell $SHELL_NIX
-
-	NEWDIR=$(cat $CAGEFILE)
-	rm $CAGEFILE
-	[ -n "$NEWDIR" ] && cd $NEWDIR
+verify_signature() {
+    local file=$1 out=
+    if out=$(gpg2 --status-fd 1 --verify shell.nix.sig shell.nix 2>/dev/null) &&
+       echo "$out" | grep -qs "^\[GNUPG:\] VALIDSIG $2 " &&
+       echo "$out" | grep -qs "^\[GNUPG:\] TRUST_ULTIMATE\$"; then
+	return 0
+    else
+	echo "$out" >&2
+	return 1
     fi
+}
+
+chpwd() {
+    if [ -n "$TRUSTED_KEY" ]; then
+      SHELL_NIX=$(findup shell.nix)
+      if [ -n "$SHELL_NIX" ]; then
+	  if [ -e "$SHELL_NIX".sig ]; then
+	      if verify_signature $SHELL_NIX.sig $TRUSTED_KEY; then
+		  echo "good sig found."
+		  # The protocol between this helper and the shell spawned is that
+		  # this shell will chdir to a dir written to file descriptor 3
+		  # by nix-shell or its subprocesses.
+		  CAGEFILE=$(mktemp /tmp/cage.XXXXXX)
+		  exec 3> $CAGEFILE
+
+		  nix-shell $SHELL_NIX
+		  NEWDIR=$(cat $CAGEFILE)
+		  rm $CAGEFILE
+		  if [ -n "$NEWDIR" ]; then
+		     cd $NEWDIR
+		  else
+		      echo "Returning to original dir."
+		      cd $OLDPWD
+		  fi
+	      else
+		  echo "signature verification failed. NOT IN SHELL".
+	      fi
+	  else
+	      echo "shell.nix found but no shell.nix.sig."
+	  fi
+      fi
+   fi
 }
